requirements.txt
aiohttp
async_generator
bs4
cowpy
cowsay
cryptg
dnspython
emoji
gitpython
google-api-python-client
google-auth-oauthlib
googletrans>=2.4.0
google_auth_oauthlib
google_images_download>=2.7.1
gTTS>=2.0.1
gTTS-token>=1.1.3
hachoir
hastebin.py>=0.2
httplib2
humanize
lmgtfy
lxml
oauth2client
Pillow>=5.3.0
psycopg2
psycopg2-binary
pybase64>=0.4.0
pyfiglet
python-dotenv
pythonping
youtube-dl
pytz
qrcode
requests>=2.18.4
speedtest-cli>=2.0.2
sqlalchemy>=1.2
telegraph
telethon>=1.10.3
telethon-session-sqlalchemy>=0.2.6
urbandict>=0.5
wikipedia>=1.4.0
pySmartDL
aria2p
search-engine-parser
